# docs/swagger.yaml
basePath: /
definitions:
  github_com_LanreAkintayo_outpost_internal_dto.ApplicationResponse:
    properties:
      api_key:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  github_com_LanreAkintayo_outpost_internal_dto.BatchReplayRequest:
    properties:
      endpoint_id:
        type: string
      from:
        type: string
      status:
        $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_models.DeliveryStatus'
      to:
        type: string
    type: object
  github_com_LanreAkintayo_outpost_internal_dto.BatchReplayResponse:
    properties:
      queued_deliveries:
        type: integer
      status:
        type: string
    type: object
  github_com_LanreAkintayo_outpost_internal_dto.CreateApplicationRequest:
    properties:
      name:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - name
    type: object
...
  github_com_LanreAkintayo_outpost_internal_dto.EventResponse:
    properties:
      application_id:
        type: string
      created_at:
        type: string
      event_type_id:
        type: string
      id:
        type: string
      idempotency_key:
        type: string
      payload:
        items:
          type: integer
        type: array
      recipient_id:
        type: string
    type: object
...
  github_com_LanreAkintayo_outpost_internal_dto.IngestEventResponse:
    properties:
      created_at:
        type: string
      event_type:
        type: string
      id:
        type: string
      idempotency_key:
        type: string
      queued_deliveries:
        type: integer
      recipient_id:
        type: string
      status:
        type: string
    type: object
...
  github_com_LanreAkintayo_outpost_internal_dto.SendEventRequest:
    properties:
      event_type:
        maxLength: 255
        minLength: 1
        type: string
      idempotency_key:
        type: string
      payload:
        items:
          type: integer
        type: array
      recipient_id:
        type: string
    required:
    - event_type
    - payload
    type: object
...
  github_com_LanreAkintayo_outpost_internal_models.DeliveryStatus:
    enum:
    - pending
    - processing
    - delivered
    - failed
    - dead_letter
    type: string
    x-enum-varnames:
    - DeliveryStatusPending
    - DeliveryStatusProcessing
    - DeliveryStatusDelivered
    - DeliveryStatusFailed
    - DeliveryStatusDeadLetter
...
  github_com_LanreAkintayo_outpost_internal_response.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
info:
  contact:
    name: Outpost Support
  description: High-performance, fault-tolerant webhook delivery platform with exponential
    retries, rate limiting, HMAC signing, and dead-letter queues.
  license:
    name: MIT
  title: Outpost Webhook Delivery Engine API
  version: "1.0"
paths:
  /api/v1/applications:
    post:
      consumes:
      - application/json
      description: Register a new tenant application and generate an API key (op_live_...)
      parameters:
      - description: Application payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_dto.CreateApplicationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_dto.ApplicationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
      summary: Create application
      tags:
      - Applications
  /api/v1/applications/{id}:
    get:
      description: Fetch application details by UUID
      parameters:
      - description: Application UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_dto.ApplicationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
      summary: Get application by ID
      tags:
      - Applications
...
  /api/v1/events:
    post:
      consumes:
      - application/json
      description: Ingest a webhook event, check idempotency, schedule delivery attempts
        to subscribed endpoints, and enqueue tasks
      parameters:
      - description: Event payload and routing parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_dto.SendEventRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_dto.IngestEventResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Ingest event
      tags:
      - Events
  /api/v1/events/{id}:
    get:
      description: Fetch details of an ingested event by UUID
      parameters:
      - description: Event UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_dto.EventResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get event by ID
      tags:
      - Events
...
  /api/v1/replay:
    post:
      consumes:
      - application/json
      description: Re-queue delivery attempts matching filter criteria (status, endpoint,
        date range)
      parameters:
      - description: Batch replay filter options
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_dto.BatchReplayRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_dto.BatchReplayResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Batch replay deliveries
      tags:
      - Deliveries
...
securityDefinitions:
  BearerAuth:
    description: Enter your API key with the Bearer prefix, e.g. 'Bearer op_live_...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"